Refactor dratsconfig.py
Refactor classes into individual modules to make future maintenance
easier.

d_rats/mainapp.py:
  Adjust for new singleton DratsConfig class.
  Convert MainApp to be a singleton class.
  Adjust for config refactor.
  Fix to better tolerate syntax errors in D-Rats config file
  for radio definitions.

d_rats/mainwindow.py:
d_rats/session_coordinator.py:
d_rats/ui/test_message_popup_model.py:
d_rats/map/mapdisplay.py:
  Adjust for refactored DratsConfig class.

d_rats/dratsconfig.py:
  Refactored from config.py and converted to singleton.

d_rats/config_defaults.py:
d_rats/configui/drats*panel.py:
d_rats/configui/config_exception.py:
d_rats/configui/dratsconfigwidget.py:
docs/source/d_rats.rst
  Refactored from config.py.

d_rats/ui/main_files.py:
  Change some noisy log messages to debug level.

docs/source/d_rats.configui.rst:
  New for the config refactor

debian_packages.sh:
  New for helping setup a Debian system for D-rats.
