+++ d_rats/configui/config_tips.py
@@ -1,5 +1,6 @@
-#!/usr/bin/python
-'''Config Tips'''
+# File: configui/config_tips.py
+
+'''Config Tips.'''
 
 if not '_' in locals():
     import gettext
@@ -10,11 +11,11 @@ TIPS_USER = {
     "latitude" : _("Your current latitude."
                    "  Use decimal degrees (DD.DDDDD)\nor D*M'S\"."
                    "  Use a space for special characters").replace("*",
-                                                                   u"\u00B0"),
+                                                                   "\u00B0"),
     "longitude" : _("Your current longitude."
                     "  Use decimal degrees (DD.DDDDD)\nor D*M'S\"."
                     "  Use a space for special characters").replace("*",
-                                                                    u"\u00B0"),
+                                                                    "\u00B0"),
     "altitude" : _("Your current altitude"),
 }
