+++ d_rats/configui/dratsoutemailpanel.py
@@ -1,0 +1,76 @@
+# File: configui/dratsoutemailpanel.py
+
+'''D-Rats Outgoing Email Panel Module.'''
+
+# Copyright 2009 Dan Smith <dsmith@danplanet.com>
+# review 2015-2020 Maurizio Andreotti  <iz2lxi@yahoo.it>
+# Copyright 2021-2024 John. E. Malmberg - Python3 Conversion
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import logging
+
+if not '_' in locals():
+    import gettext
+    _ = gettext.gettext
+
+from .dratspanel import DratsPanel
+from .dratspanel import disable_with_toggle
+from .dratsconfigwidget import DratsConfigWidget
+
+
+class DratsOutEmailPanel(DratsPanel):
+    '''
+    D-Rats Out Email Panel.
+
+    :param dialog: D-Rats Config UI Dialog
+    :type dialog: :class:`config.DratsConfigUI`
+    '''
+    logger = logging.getLogger("DratsOutEmailPanel")
+
+    # pylint: disable=unused-argument
+    def __init__(self, dialog=None):
+        DratsPanel.__init__(self)
+
+        gateway = DratsConfigWidget(section="settings", name="smtp_dogw")
+        gateway.add_bool()
+        self.make_view(_("SMTP Gateway"), gateway)
+
+        val = DratsConfigWidget(section="settings", name="smtp_server")
+        val.add_text()
+        self.make_view(_("SMTP Server"), val)
+        disable_with_toggle(gateway.child_widget, val.child_widget)
+
+        port = DratsConfigWidget(section="settings", name="smtp_port")
+        port.add_numeric(1, 65536, 1)
+        mode = DratsConfigWidget(section="settings", name="smtp_tls")
+        mode.add_bool("TLS")
+        self.make_view(_("Port and Mode"), port, mode)
+        disable_with_toggle(gateway.child_widget, port.child_widget)
+        disable_with_toggle(gateway.child_widget, mode.child_widget)
+
+        val = DratsConfigWidget(section="settings", name="smtp_replyto")
+        val.add_text()
+        self.make_view(_("Source Address"), val)
+        disable_with_toggle(gateway.child_widget, val.child_widget)
+
+        val = DratsConfigWidget(section="settings", name="smtp_username")
+        val.add_text()
+        self.make_view(_("SMTP Username"), val)
+        disable_with_toggle(gateway.child_widget, val.child_widget)
+
+        val = DratsConfigWidget(section="settings", name="smtp_password")
+        val.add_pass()
+        self.make_view(_("SMTP Password"), val)
+        disable_with_toggle(gateway.child_widget, val.child_widget)
