+++ docs/source/d_rats.configui.rst
@@ -1,0 +1,197 @@
+d\_rats.configui package
+========================
+
+Submodules
+----------
+
+d\_rats.configui.config\_exception module
+-----------------------------------------
+
+.. automodule:: d_rats.configui.config_exception
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.config\_tips module
+------------------------------------
+
+.. automodule:: d_rats.configui.config_tips
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsappearancepanel module
+--------------------------------------------
+
+.. automodule:: d_rats.configui.dratsappearancepanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratschatpanel module
+--------------------------------------
+
+.. automodule:: d_rats.configui.dratschatpanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsconfigui module
+-------------------------------------
+
+.. automodule:: d_rats.configui.dratsconfigui
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsconfigwidget module
+-----------------------------------------
+
+.. automodule:: d_rats.configui.dratsconfigwidget
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsemailaccesspanel module
+---------------------------------------------
+
+.. automodule:: d_rats.configui.dratsemailaccesspanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsgpsexportpanel module
+-------------------------------------------
+
+.. automodule:: d_rats.configui.dratsgpsexportpanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsgpspanel module
+-------------------------------------
+
+.. automodule:: d_rats.configui.dratsgpspanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsinemailpanel module
+-----------------------------------------
+
+.. automodule:: d_rats.configui.dratsinemailpanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratslistconfigwidget module
+---------------------------------------------
+
+.. automodule:: d_rats.configui.dratslistconfigwidget
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsmappanel module
+-------------------------------------
+
+.. automodule:: d_rats.configui.dratsmappanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsmessagepanel module
+-----------------------------------------
+
+.. automodule:: d_rats.configui.dratsmessagepanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsnetworkpanel module
+-----------------------------------------
+
+.. automodule:: d_rats.configui.dratsnetworkpanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsoutemailpanel module
+------------------------------------------
+
+.. automodule:: d_rats.configui.dratsoutemailpanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratspanel module
+--------------------------------------
+
+.. automodule:: d_rats.configui.dratspanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratspathspanel module
+---------------------------------------
+
+.. automodule:: d_rats.configui.dratspathspanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsprefspanel module
+---------------------------------------
+
+.. automodule:: d_rats.configui.dratsprefspanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratsradiopanel module
+---------------------------------------
+
+.. automodule:: d_rats.configui.dratsradiopanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratssoundpanel module
+---------------------------------------
+
+.. automodule:: d_rats.configui.dratssoundpanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratstcpincomingpanel module
+---------------------------------------------
+
+.. automodule:: d_rats.configui.dratstcpincomingpanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratstcpoutgoingpanel module
+---------------------------------------------
+
+.. automodule:: d_rats.configui.dratstcpoutgoingpanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratstcppanel module
+-------------------------------------
+
+.. automodule:: d_rats.configui.dratstcppanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+d\_rats.configui.dratstransferspanel module
+-------------------------------------------
+
+.. automodule:: d_rats.configui.dratstransferspanel
+    :members:
+    :undoc-members:
+    :show-inheritance:
